refactor: move selfieUrl function to dedicate export
